+++ plugins/inputs/socket_listener/socket_listener.go
@@ -10,6 +10,7 @@ import (
 	"sync"
 
 	"github.com/influxdata/telegraf"
+	"github.com/influxdata/telegraf/internal"
 	"github.com/influxdata/telegraf/plugins/inputs"
 	"github.com/influxdata/telegraf/plugins/parsers"
 	"time"
@@ -76,8 +77,8 @@ func (ssl *streamSocketListener) read(c net.Conn) {
 		for _, m := range metrics {
 			ssl.AddFields(m.Name(), m.Fields(), m.Tags(), m.Time())
 		}
-		if ssl.ReadDeadline > 0 {
-			c.SetReadDeadline(time.Now().Add(time.Duration(ssl.ReadDeadline) * time.Second))
+		if ssl.ReadTimeout.Duration > 0 {
+			c.SetReadDeadline(time.Now().Add(ssl.ReadTimeout.Duration * time.Second))
 		}
 	}
 
@@ -116,7 +117,7 @@ type SocketListener struct {
 	ServiceAddress string
 	MaxConnections int
 	ReadBufferSize int
-	ReadDeadline   int
+	ReadTimeout    internal.Duration
 
 	parsers.Parser
 	telegraf.Accumulator